Fixed issue with greyed out operations on metadata page for Turkish users
When the client culture is set to tr-TR (Turkish) and the ASP.NET client
based culture settings are configured as follows in web.config:

<system.web>
    ...
    <globalization uiCulture="auto" culture="auto"
                   enableClientBasedCulture="true" />
    ...
</system.web>

Operations under the metadata page which contain a capital 'I' will be
grayed out. This is known as the Turkish 'I' problem.

The problem is using ToLower() with strings containing a capital 'I' when
the current culture is tr-TR (Turkish)

The problem is described in detail here:
http://haacked.com/archive/2012/07/05/turkish-i-problem-and-why-you-should-care.aspx

and here:
http://www.codinghorror.com/blog/2008/03/whats-wrong-with-turkey.html